✏️🚚 Change: rename and move `stream` module to `param::news_stream` [#63]
