+++ src/client/param/mod.rs
@@ -1,0 +1,3 @@
+//! Features for the parameters of the [`Client`](crate::client::Client) struct methods.
+
+pub mod news_stream;
