+++ tests/client.rs
@@ -1,4 +1,4 @@
-use tetr_ch::client::*;
+use tetr_ch::client::{*, param::news_stream::*};
 
 #[test]
 fn get_usr_data() {
@@ -18,13 +18,13 @@ fn get_server_activity_data() {
 
 #[test]
 fn get_latest_global_news_data() {
-    let _ = tokio_test::block_on(Client::new().get_news_latest(stream::NewsStream::Global, 3));
+    let _ = tokio_test::block_on(Client::new().get_news_latest(NewsStream::Global, 3));
 }
 
 #[test]
 fn get_latest_user_scale_news_data() {
     let _ = tokio_test::block_on(Client::new().get_news_latest(
-        stream::NewsStream::User("621db46d1d638ea850be2aa0".to_string()),
+        NewsStream::User("621db46d1d638ea850be2aa0".to_string()),
         3,
     ));
 }
@@ -32,5 +32,5 @@ fn get_latest_user_scale_news_data() {
 #[test]
 #[should_panic]
 fn panic_if_invalid_limit_range_in_getting_latest_news() {
-    let _ = tokio_test::block_on(Client::new().get_news_latest(stream::NewsStream::Global, 0));
+    let _ = tokio_test::block_on(Client::new().get_news_latest(NewsStream::Global, 0));
 }
